+++ cloudsmith_cli/core/api/user.py
@@ -97,3 +97,4 @@ def refresh_user_token(token_slug: str) -> dict:
 
     ratelimits.maybe_rate_limit(client, headers)
     return data
+